refactor: oauth 개발 전 폴더 구조 리팩토링
