+++ src/main/java/org/sophy/sophy/config/auth/SecurityConfig.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.config;
+package org.sophy.sophy.config.auth;
 
 import java.util.Arrays;
 import java.util.List;
@@ -12,6 +12,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -42,7 +43,7 @@ public class SecurityConfig {
         //CSRF 설정 Disable
         http.csrf().disable()
 
-            .cors(cors -> cors.disable())
+            .cors(AbstractHttpConfigurer::disable)
 
             //exception handling 할 때 우리가 만든 클래스를 추가
             .exceptionHandling()
